Rework amount storage
amount are now stored as integer (ie in cents) except for beneficiary
where full precision is kept until debt creation.

Please update your db with update-2012-04-10.sql

+++ inc/class/Eu/Rmmt/Debt/Debt.class.php
@@ -49,7 +49,7 @@ class Debt
     private $_amount;
 
     public function __construct(User $from, User $to, $amount) {
-        $amount = (float)$amount;
+        $amount = (int)$amount;
 
         if ($amount > 0) {
             $this->_from   = $from;
@@ -91,7 +91,7 @@ class Debt
 
     public function setAmount($amount)
     {
-        $amount = (float)$amount;
+        $amount = (int)$amount;
         if ($amount > 0) {
             $this->_amount = $amount;
         } elseif ($amount < 0) {
